new updates to project structure and conditional rendering modified
